score相关的controller
还修改了honor中的一处错误

+++ java-fx/src/main/java/com/teach/javafx/models/DO/Score.java
@@ -35,7 +35,7 @@ public class Score implements Serializable {
     @JoinColumn(name = "course_id")
     private Course course;
 
-    private Integer mark;
+    private Double mark;
 
     private Integer ranking;
